requirements.txt
aiohttp
gidgethub>=4.1.0
PyJWT
cryptography
cachetools
gitpython
mypy
